build-requirements.txt
PyInstaller>=6.16.0
Pillow>=11.0.0
